# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Boutique Commodity Shop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      responses:
        "200":
          description: Wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
    post:
      operationId: addToWishlist
      tags: [wishlist]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
  /wishlist/{productId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ─── ORDERS ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, price, categoryId, unit, stockQuantity, minStockLevel, status]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        unit:
          type: string
        stockQuantity:
          type: integer
        minStockLevel:
          type: integer
        imageUrl:
          type: ["string", "null"]
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        status:
          type: string
          enum: [in_stock, low_stock, out_of_stock]
        rating:
          type: number
        reviewCount:
          type: integer
        isFeatured:
          type: boolean
        discountPercent:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    WishlistItem:
      type: object
      required: [id, productId, product]
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
    WishlistInput:
      type: object
      required: [productId]
      properties:
        productId:
          type: integer

    # ─── ORDER ────────────────────────────────────────────────────────────────
...